Achieved Function: Redis : User profile showing followee and follower

+++ Community/src/main/java/com/robin/community/service/LikeService.java
@@ -7,14 +7,14 @@ import org.springframework.data.redis.core.RedisOperations;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.core.SessionCallback;
 import org.springframework.stereotype.Service;
-
 @Service
 public class LikeService {
+
     @Autowired
     private RedisTemplate redisTemplate;
 
     // 点赞
-    public void like(int userId, int entityType, int entityId, int entityUserId){
+    public void like(int userId, int entityType, int entityId, int entityUserId) {
         redisTemplate.execute(new SessionCallback() {
             @Override
             public Object execute(RedisOperations operations) throws DataAccessException {
@@ -37,6 +37,7 @@ public class LikeService {
             }
         });
     }
+
     // 查询某实体点赞的数量
     public long findEntityLikeCount(int entityType, int entityId) {
         String entityLikeKey = RedisKeyUtil.getEntityLikeKey(entityType, entityId);
@@ -53,6 +54,7 @@ public class LikeService {
     public int findUserLikeCount(int userId) {
         String userLikeKey = RedisKeyUtil.getUserLikeKey(userId);
         Integer count = (Integer) redisTemplate.opsForValue().get(userLikeKey);
-        return count == null ? 0 : count;
+        return count == null ? 0 : count.intValue();
     }
+
 }
